==> notebook/1_Webscraping_IGN.ipynb <==
import pandas as pd
%run 5_Funciones.ipynb
from selenium.webdriver.chrome.options import Options
from selenium import webdriver
from selenium.webdriver.common.by import By
import time

En primer lugar, voy a configurar un WebDriver para Selenium con opciones específicas, incluyendo deshabilitar la automatización para evitar la detección de bots, configurar el navegador como "headless" (invisible), maximizar la ventana y abrir una ventana de incógnito, utilizando el controlador de Chrome ubicado en 'drive/chromedriver':

opciones = Options()
opciones.add_experimental_option('excludeSwitches', ['enable-automation'])
opciones.add_experimental_option('useAutomationExtension', False)
opciones.headless = False
opciones.add_argument('--start-maximized')
opciones.add_argument('--incognito')

PATH = 'drive/chromedriver'
driver = webdriver.Chrome(options=opciones)

A continuación, voy a utilizar el WebDriver para abrir y cargar la página web que me interesa:

url = "https://www.ign.es/web/ane-datos-geograficos/-/datos-geograficos/datosPoblacion?tipoBusqueda=capitales"
driver.get(url)

'''
Utilizo Selenium para localizar y almacenar en la variable 
tabla el elemento HTML de la tabla específica en la página web.
'''
tabla = driver.find_element(By.XPATH, '//*[@id="p_p_id_IGNANEDatosGeograficos_WAR_IGNANEDatosGeograficosportlet_"]/div/div/div/div/table[2]')
tabla

'''
Localizo y obtengo todas las filas (<tr>) dentro del 
cuerpo (<tbody>) de la tabla utilizando Selenium.
'''
cuerpo = tabla.find_element(By.TAG_NAME, 'tbody')

filas = cuerpo.find_elements(By.TAG_NAME, 'tr')

'''
Itero sobre las filas de una tabla, extrayendo el 
texto de las celdas (<td>) en cada fila y almacenando 
esos datos en una lista 'datos_lista'
'''
datos_filas = []
for fila in filas:
    celdas = fila.find_elements(By.TAG_NAME, 'td')
    datos_fila = [celda.text for celda in celdas]
    datos_filas.append(datos_fila)

Una vez realizada la extracción, creo un DataFrame de Pandas llamado data utilizando la información extraída de las filas de la tabla y asignando nombres de columna específicos:

columnas = ['Capital', 'Provincia', 'Población', 'Latitud', 'Longitud', 'XUTM ETRS89', 'YUTM ETRS89', 
            'Altura (m)', 'Huso']
data = pd.DataFrame(datos_filas, columns=columnas)

data.dtypes

Acontinuación, elimino las columnas y filas (Ceuta y Melilla) que no me interesan para el análisis:

data= data.drop(columns=['XUTM ETRS89', 'YUTM ETRS89', 'Altura (m)', 'Capital', 'Huso'])

filas_eliminar= [0,13,30]
data= data.drop(filas_eliminar).reset_index(drop=True)
data.head()

Realizaré una corrección de nombres de provincias en la columna 'Provincia' del DataFrame 'data' mediante un diccionario de mapeo para que en todos los CSV estén estandarizados:

correccion_provincias={'Álava': 'Araba/alava',
                      'Santa Cruz de Tenerife':'Sta. cruz de tenerife',
                      'Málaga':'Malaga',
                      'Las Palmas': 'Las palmas de gran canaria',
                      'Valencia / València':'Valencia',
                      'La Rioja': 'La rioja',
                      'León':'Leon',
                      'Alicante / Alacant':'Alicante',
                      'Almería':'Almeria',
                      'Ávila': 'Avila',
                      'Illes Balears': 'Islas baleares',
                      'Cáceres': 'Caceres',
                      'Cádiz':'Cadiz',
                      'Castellón / Castelló': 'Castellon',
                      'Ciudad Real': 'Ciudad real',
                      'Córdoba': 'Cordoba',
                      'A Coruña': 'A coruna',
                      'Jaén': 'Jaen' }

data['Provincia']= data['Provincia'].replace(correccion_provincias)

Trato los tipos de datos:

data['Población'] = pd.to_numeric(data['Población'].str.replace('.', ''), errors='coerce').astype(int)
data['Latitud'] = pd.to_numeric(data['Latitud'].str.replace(',', '.'), errors='coerce')
data['Longitud'] = pd.to_numeric(data['Longitud'].str.replace(',', '.'), errors='coerce')

data.dtypes

data.to_csv('../data/IGN.csv', index=False)
