==> notebook/2_AEMET.ipynb <==
import pandas as pd
import numpy as np
import warnings
warnings.filterwarnings("ignore")
import requests
import json
import os
import locale
import requests
%run 5_Funciones.ipynb
pd.set_option('display.max_columns', None) 
pd.set_option('display.max_rows', None)
from datetime import datetime, timedelta

Primero, llamos a la función 'obtener_encabezado_api' del JN 4_Funciones, para obtener el encabezado de autorización AEMET utilizando un token almacenado en el archivo '../api.txt':

aemet = obtener_encabezado_api(api_file_path='../api.txt')

Se definen las columnas de interés para el análisis de datos climáticos:

columnas_interes= ['fecha', 'provincia', 'tmed', 'prec', 'sol', 'tmin', 'tmax']

columnas_procesar= ['tmed', 'prec', 'sol', 'tmin', 'tmax']

'''
Defino la lista de años a procesar (desde 2014 hasta 2023)
'''
anos_a_procesar = list(range(2014, 2024))

lista_dataframes = []

'''
Creo la lista vacía 'lista_dataframes' y la lleno iterando a través de años y meses, 
generando y procesando DataFrames mensuales con la función 'procesar_datos_mes', si las 
URLs de datos son válidas.
'''

for ano in anos_a_procesar:
    
    for mes in range(1, 13):
        
        fecha_inicio = f'{ano}-{mes:02d}-01T00:00:00UTC'
        fecha_fin = f'{ano}-{mes:02d}-31T00:00:00UTC'
        url_datos = obtener_url_datos(fecha_inicio, fecha_fin)

        if url_datos:
            df_mes = procesar_datos_mes(url_datos, columnas_interes)
            lista_dataframes.append(df_mes)

Una vez obtenido los datos de la API y procesados, el siguiente paso es concatener todos los DataFrames en uno solo:

df_final = pd.concat(lista_dataframes, ignore_index=True)

data = df_final.copy()

data['Provincia'].value_counts()

A continuación realizaré una corrección de nombres de provincias en la columna 'Provincia' del DataFrame 'data' mediante un diccionario de mapeo:

correccion_provincias = {'A coruña': 'A coruna',
                         'Las palmas':'Las palmas de gran canaria',
                         'Illes balears' : 'Islas baleares',
                        'Bizkaia': 'Vizcaya'}

data['Provincia'].replace(correccion_provincias, inplace=True)
data.head()

Ya casi tengo el DataFrame de clima preparado, unicamente me falta añadir la latitud y longitud de cada provincia para poder hacer la representación visual correspondiente. Por lo que, una vez tratados los datos en el JP 'Webscraping_IGN' y exportado en CSV, lo llamo a continuación y realizo un join inner por Provincia:

lat_long= pd.read_csv('../data/IGN.csv')

df_merged = pd.merge(data, lat_long, on='Provincia', how='inner')
df_merged.head()

df_merged['Periodo'] = df_merged['Periodo'].dt.to_timestamp()

df_merged.dtypes

df_merged.shape

df_merged.columns

Por último, guardo el df_merged como clima:

df_merged.to_csv('../data/clima.csv', index=False)
