==> notebook/4_SQL.ipynb <==
# SQL

Una vez realizada toda la limpieza y ordenado todos mis datos. Me conecto con la base de datos a SQL para establecer relaciones:

import pandas as pd
from sqlalchemy import create_engine
%run 5_Funciones.ipynb

usuario = 'root'
ruta_archivo_contraseña = '../SQL_pass.txt'  # Reemplaza con la ruta correcta al archivo de texto
host = 'localhost'
puerto = '3306'
nombre_bd = 'turismo'

cadena_conexion = crear_cadena_conexion(usuario, ruta_archivo_contraseña, host, puerto)
crear_base_de_datos(cadena_conexion, nombre_bd)


#### INE

ruta_csv_ine_viajeros = '../data/ine_viajeros.csv'
nombre_tabla_mysql = 'viajeros'
cargar_csv_a_mysql(ruta_csv_ine_viajeros, cadena_conexion, nombre_tabla_mysql, nombre_bd)

ruta_csv_ine_pernoctaciones = '../data/ine_pernoctaciones.csv'
nombre_tabla_mysql = 'pernoctaciones'
cargar_csv_a_mysql(ruta_csv_ine_pernoctaciones, cadena_conexion, nombre_tabla_mysql, nombre_bd)

ruta_csv_ine_total = '../data/ine_total.csv'
nombre_tabla_mysql = 'total'
cargar_csv_a_mysql(ruta_csv_ine_total, cadena_conexion, nombre_tabla_mysql, nombre_bd)

AEMET

ruta_csv_aemet_meses = '../data/clima.csv'
nombre_tabla_mysql = 'clima'
cargar_csv_a_mysql(ruta_csv_aemet_meses, cadena_conexion, nombre_tabla_mysql, nombre_bd)
